+++ RELEASE_NOTES.md
@@ -31,6 +31,12 @@ Logtalk compiler and runtime
 * ADDED: New `fails_if` key to the `info/2` predicate directive for documenting
 predicate fail conditions. Adopted from ECLiPSe similar `fail_if` key.
 
+* IMPROVED: Detect and report when a `uses/2` directive references the object
+that contains it.
+
+* IMPROVED: Detect and report when a `use_module/2` directive references the
+module (being compiled as an object) that contains it.
+
 * IMPROVED: Detect and report when a source file attempts to load itself.
 
 * IMPROVED: Error reporting when a source file attempts to include itself.